# docs/openapi-snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jekko-server",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/config": {
      "get": {
        "tags": [
          "crate::routes::config"
        ],
        "summary": "`GET /api/v1/config` — current config snapshot.",
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "Current config"
          }
        }
      },
      "put": {
        "tags": [
          "crate::routes::config"
        ],
        "summary": "`PUT /api/v1/config` — replace the in-memory config.",
        "operationId": "update_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated config"
          }
        }
      }
    },
...
    "/api/v1/daemon/preview": {
      "post": {
        "tags": [
          "crate::routes::daemon"
        ],
        "summary": "`POST /api/v1/daemon/preview` — synthesise a preview without executing.",
        "operationId": "preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Preview"
          }
        }
      }
    },
    "/api/v1/daemon/{run_id}": {
      "get": {
        "tags": [
          "crate::routes::daemon"
        ],
        "summary": "`GET /api/v1/daemon/:run_id`.",
        "operationId": "get_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daemon run"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v1/daemon/{run_id}/abort": {
      "post": {
        "tags": [
          "crate::routes::daemon"
        ],
        "summary": "`POST /api/v1/daemon/:run_id/abort`.",
        "operationId": "abort",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aborted"
          }
        }
      }
    },
    "/api/v1/daemon/{run_id}/pause": {
      "post": {
        "tags": [
          "crate::routes::daemon"
        ],
        "summary": "`POST /api/v1/daemon/:run_id/pause`.",
        "operationId": "pause",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paused"
          }
        }
      }
    },
    "/api/v1/daemon/{run_id}/resume": {
      "post": {
        "tags": [
          "crate::routes::daemon"
        ],
        "summary": "`POST /api/v1/daemon/:run_id/resume`.",
        "operationId": "resume",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumed"
          }
        }
      }
    },
...
    "/api/v1/experimental/{flag}": {
      "get": {
        "tags": [
          "crate::routes::experimental"
        ],
        "summary": "`GET /api/v1/experimental/:flag`.",
        "operationId": "get_flag",
        "parameters": [
          {
            "name": "flag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flag"
          }
        }
      },
      "put": {
        "tags": [
          "crate::routes::experimental"
        ],
        "summary": "`PUT /api/v1/experimental/:flag`.",
        "operationId": "set_flag",
        "parameters": [
          {
            "name": "flag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      }
    },
...
    "/api/v1/mcp/{name}": {
      "get": {
        "tags": [
          "crate::routes::mcp"
        ],
        "summary": "`GET /api/v1/mcp/:name` — fetch one server.",
        "operationId": "get_server",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "MCP server"
          }
        }
      }
    },
...
    "/api/v1/permission/{id}": {
      "post": {
        "tags": [
          "crate::routes::permission"
        ],
        "summary": "`POST /api/v1/permission/:id` — reply to a pending ask.",
        "operationId": "reply",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replied"
          }
        }
      }
    },
...
    "/api/v1/provider/{id}/authorize": {
      "get": {
        "tags": [
          "crate::routes::provider"
        ],
        "summary": "`GET /api/v1/provider/:id/authorize`.",
        "operationId": "authorize_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorize response"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FlagBody": {
        "type": "object",
        "description": "Body of `PUT /api/v1/experimental/:flag`.",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "object",
            "description": "New value."
          }
        }
      },
...
      "PreviewBody": {
        "type": "object",
        "description": "Request body for `POST /api/v1/daemon/preview`.",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string",
            "description": "Free-form text to preview."
          }
        }
      },
...
      "ReplyBody": {
        "type": "object",
        "description": "Body of `POST /api/v1/permission/:id`. The inner reply is exchanged as a\nstring (`\"once\"` / `\"always\"` / `\"reject\"`) so the schema does not\ndepend on `PermissionReply: ToSchema`.",
        "required": [
          "reply"
        ],
        "properties": {
          "reply": {
            "type": "string",
            "description": "`once` / `always` / `reject`."
          }
        }
      },
...
    }
  }
}